% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{check.ibs}
\alias{check.ibs}
\title{Check Identity-By-State (IBS) for a genotype pair}
\usage{
check.ibs(gen)
}
\arguments{
\item{gen}{Numeric vector of length two with genotype codes.}
}
\value{
Integer: 2 if identical non-heterozygotes, 0 if opposite homozygotes, -1 otherwise.
}
\description{
Checks IBS status for two genotypes.
}
\examples{
check.ibs(c(1, 1))
check.ibs(c(1, 3))

}
