% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{check.identical.samples.by.block}
\alias{check.identical.samples.by.block}
\title{Check identical samples by block}
\usage{
check.identical.samples.by.block(genotypes, blcsize, threshold = 0)
}
\arguments{
\item{genotypes}{Genotype matrix.}

\item{blcsize}{Block size (number of SNPs).}

\item{threshold}{Distance threshold. Default 0.}
}
\value{
List of identical sample pairs.
}
\description{
Identifies identical samples within SNP blocks.
}
\examples{
# See check.identical.samples example

}
