% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{check.mendelian.inconsistencies.pair}
\alias{check.mendelian.inconsistencies.pair}
\title{Check Mendelian inconsistencies for a pair}
\usage{
check.mendelian.inconsistencies.pair(g1, g2)
}
\arguments{
\item{g1}{Genotype vector for parent.}

\item{g2}{Genotype vector for child.}
}
\value{
Numeric vector: [# inconsistencies, # comparable SNPs].
}
\description{
Calculates number of inconsistencies and total comparable SNPs for a parent-child pair.
}
\examples{
# Used internally by check.mendelian.inconsistencies

}
