% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{check.mendelian.inconsistencies}
\alias{check.mendelian.inconsistencies}
\title{Check Mendelian inconsistencies}
\usage{
check.mendelian.inconsistencies(genotypes, father, child)
}
\arguments{
\item{genotypes}{Genotype matrix.}

\item{father}{Vector of father sample IDs.}

\item{child}{Vector of child sample IDs.}
}
\value{
Data frame summarizing inconsistencies per pair.
}
\description{
Identifies Mendelian inconsistencies between father-child pairs.
}
\examples{
# Requires proper parent-child genotype data

}
