% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSamples.R
\name{check.sample.call.rate}
\alias{check.sample.call.rate}
\title{Check Sample Call Rate}
\usage{
check.sample.call.rate(sample.summary, min.call.rate)
}
\arguments{
\item{sample.summary}{A data frame with a "Call.rate" column for each sample.}

\item{min.call.rate}{Minimum acceptable call rate (between 0 and 1).}
}
\value{
A character vector with the names of samples to remove.
}
\description{
Identifies samples with call rate below a given threshold.
}
