% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{check.sample.heterozygosity}
\alias{check.sample.heterozygosity}
\title{Check sample heterozygosity}
\usage{
check.sample.heterozygosity(sample.summary, max.dev)
}
\arguments{
\item{sample.summary}{Data frame containing sample summary (must have `Heterozygosity` column).}

\item{max.dev}{Maximum number of standard deviations allowed from mean.}
}
\value{
Character vector with sample names considered outliers. Returns `NULL` if none.
}
\description{
Identifies samples with heterozygosity values deviating beyond a specified threshold.
}
\examples{
ss <- data.frame(Heterozygosity = c(0.2, 0.5, 0.7))
rownames(ss) <- c("Ind1", "Ind2", "Ind3")
check.sample.heterozygosity(ss, 1)

}
