% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{check.snp.chromo}
\alias{check.snp.chromo}
\title{Check SNP by chromosome}
\usage{
check.snp.chromo(snpmap, chromosomes)
}
\arguments{
\item{snpmap}{Data frame with SNP map info (must contain columns `Chromosome` and `Name`).}

\item{chromosomes}{Vector of chromosome identifiers to filter.}
}
\value{
Character vector with SNP names.
}
\description{
Filters SNP names belonging to specified chromosomes.
}
\examples{
snpmap <- data.frame(Chromosome = c(1, 1, 2), Name = c("SNP1", "SNP2", "SNP3"))
check.snp.chromo(snpmap, 1)

}
\author{
Roberto Higa
}
