% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{check.snp.monomorf}
\alias{check.snp.monomorf}
\title{Check SNP monomorphic status}
\usage{
check.snp.monomorf(snp.summary)
}
\arguments{
\item{snp.summary}{Data frame with columns `P.AA`, `P.AB`, `P.BB`.}
}
\value{
Character vector with monomorphic SNP names. Returns `NULL` if none.
}
\description{
Identifies SNPs considered monomorphic.
}
\examples{
df <- data.frame(P.AA = c(1, 0.5), P.AB = c(0, 0.5), P.BB = c(0, 0))
rownames(df) <- c("SNP1", "SNP2")
check.snp.monomorf(df)

}
\author{
Roberto Higa
}
