% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinar.R
\name{combinarSNPData}
\alias{combinarSNPData}
\title{Combina múltiplos objetos SNPDataLong (painéis diferentes)}
\usage{
combinarSNPData(lista)
}
\arguments{
\item{lista}{Lista de objetos da classe SNPDataLong}
}
\value{
Objeto SNPDataLong unificado
}
\description{
Combina múltiplos objetos SNPDataLong (painéis diferentes)
}
