% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_geno.R
\name{count_geno}
\alias{count_geno}
\title{Counts dosages of alleles}
\usage{
count_geno(geno, count = "B")
}
\arguments{
\item{geno}{matrix of genotypes with SNPs in columns and samples in rows}

\item{count}{character ("A" or "B") for which allele will be counted}
}
\value{
a new genotype matrix with counts of a particular allele rather than A/B coding
}
\description{
Counts the dosage of a particular allele when genotyping data is in A/B format
}

