% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_snp_info.R
\name{create_snp_info}
\alias{create_snp_info}
\title{Create snp_info file required as FImpute imput}
\usage{
create_snp_info(map = map, geno = genotype, gpData = NULL,
  chip_number = 1)
}
\arguments{
\item{map}{data.frame containing SNP map}

\item{geno}{data.frame or matrix containing genotypes (animals in rows, SNPs in columns)}
}
\value{
object nearly ready to be written to a table and used as FImpute input as the snp_info file
}
\description{
Author: Sebastian Casiro
}

