% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{doPCA}
\alias{doPCA}
\title{Do genome relationship matrix PCA}
\usage{
doPCA(genotypes)
}
\arguments{
\item{genotypes}{Genotype matrix.}
}
\value{
List containing `pcs` (principal components) and `eigen` (eigenvalues).
}
\description{
Performs PCA using the genome relationship matrix (GRM).
}
\examples{
# Requires matrix of numeric genotypes

}
