% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{exploratory.plots}
\alias{exploratory.plots}
\title{Exploratory plots for SNP and sample summary}
\usage{
exploratory.plots(
  snp.summary,
  snps.plot,
  sample.summary,
  samples.plot,
  distm,
  glabels,
  mds.plot,
  hierq.plot
)
}
\arguments{
\item{snp.summary}{Data frame with SNP summary.}

\item{snps.plot}{Filename for SNP histogram plot.}

\item{sample.summary}{Data frame with sample summary.}

\item{samples.plot}{Filename for heterozygosity plot.}

\item{distm}{Distance matrix for samples.}

\item{glabels}{Sample labels for plots.}

\item{mds.plot}{Filename for MDS plot.}

\item{hierq.plot}{Filename for hierarchical cluster plot.}
}
\value{
None. Plots are saved as JPEG files.
}
\description{
Generates exploratory plots: MAF histograms, HWE plots, heterozygosity scatter, MDS, and dendrogram.
}
\examples{
# Requires proper SNP and sample summary data frames

}
