% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_geno.R
\name{filter_geno}
\alias{filter_geno}
\title{Filters genotypes based on various criteria}
\usage{
filter_geno(geno, call_rate = 0.9, fix = TRUE)
}
\arguments{
\item{geno}{matrix of genotyping data with individuals in rows and SNPs in columns.
Formatted in dosage of B format.}

\item{call_rate}{numeric proportion indicating the call rate that must be met}

\item{fix}{boolean. If TRUE, remove SNPs that are fixed among all animals}
}
\value{
geno matrix with removed (filtered out) SNPs
}
\description{
Before using genotypes in models, this function can be used to filter out SNPs that are
 fixed, have a low call rate, etc.
}

