% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fimpute_run.R
\name{fimpute_run}
\alias{fimpute_run}
\title{Run Fimpute}
\usage{
fimpute_run(geno, map, ped = NULL, path = NULL)
}
\arguments{
\item{geno}{a matrix or data.frame with SNPs in columns and individuals in rows. Genotypes must
be coded as dosage of allele 'b' {0, 1, 2}.}

\item{map}{a data.frame containing SNP map information for each SNP present in geno}

\item{ped}{a data.frame pedigree providing family information for each individual in geno. The first
column of the pedigree is for ID, second is for sire/father ID, and third is for dam/mother ID.}

\item{path}{a character represting the path to the FImpute binary. If omitted, assumes FImpute binary
resides along PATH.}
}
\description{
Invoke Fimpute for imputation and genotype phasing from within R. Uses R objects for genotypes,
maps, and pedigrees to construct Fimpute inputs, and executes an Fimpute run. Fimpute is required
to be on your PATH, or otherwise specify the location of the Fimpute binary with the path argument
}

