% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FImputeExport-class}
\alias{FImputeExport-class}
\title{FImputeExport Class}
\description{
A class to handle export preparation for FImpute.
}
\section{Slots}{

\describe{
\item{\code{geno}}{A SnpMatrix or NULL containing genotype data.}

\item{\code{map}}{A data.frame containing marker information.}

\item{\code{path}}{Output file path.}

\item{\code{name}}{Project or file name.}
}}

