% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FImputeRunner-class}
\alias{FImputeRunner-class}
\title{FImputeRunner Class}
\description{
A class to manage FImpute execution and results.
}
\section{Slots}{

\describe{
\item{\code{export}}{An FImputeExport object.}

\item{\code{par_file}}{Path to parameter file.}

\item{\code{exec_path}}{Path to FImpute executable.}

\item{\code{results}}{A data.frame containing results or summary information.}
}}

