% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fimpute_runner.R
\name{FImputeRunner}
\alias{FImputeRunner}
\title{Build FImputeRunner object}
\usage{
FImputeRunner(object, path, exec_path = "FImpute3", name = "data")
}
\arguments{
\item{object}{An object of class `SNPDataLong`, from which `geno` and `map` slots will be extracted.}

\item{path}{A character string indicating the directory to save FImpute files.}

\item{exec_path}{Path to the FImpute executable (default = "FImpute3").}

\item{name}{Name for the dataset (used internally, default = "data").}
}
\value{
An object of class `FImputeRunner`.
}
\description{
A convenience function to construct a `FImputeRunner` object from a `SNPDataLong` object.
}
