% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_gpGeno.R
\name{format_gpGeno}
\alias{format_gpGeno}
\title{Synbreed geno formatting converter}
\usage{
format_gpGeno(data_file, num_ids, num_snps, head = 10, snp_col = 1,
  id_col = 2, allele1_col = 3, allele2_col = 4, format = "doseB")
}
\arguments{
\item{data_file}{Path to raw genotyping file in long format}

\item{num_ids}{Numeric. Number of individuals in file}

\item{num_snps}{Numeric. Number of SNPs per individual}

\item{head}{Numeric. Number of header lines.}

\item{snp_col}{Numeric. Indicates which column in file contains SNP names}

\item{id_col}{Numeric. Indicates which column in file contains individual names}

\item{allele1_col}{Numeric. Indicates which column contains genotyping for allele 1.}

\item{allele2_col}{Numeric. Indicates which column contains genotyping for allele 2.}

\item{format}{character describing desired output. Options include "doseB", indicating that the
 output will be in the dosage of allele B format. If doseB is not chosen, output is left in
 original format.}
}
\value{
Genotype matrix with individuals in rows and snps in columns
}
\description{
Function to convert genotype file from long format to matrix format suitable for Synbreeds
gpData object.
}

