% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{get.correl.fc}
\alias{get.correl.fc}
\title{Get correlation (fc method)}
\usage{
get.correl.fc(g1, g2)
}
\arguments{
\item{g1}{Genotype vector.}

\item{g2}{Genotype vector.}
}
\value{
Numeric value of correlation.
}
\description{
Calculates genotype correlation using a fast check (fc) method.
}
\examples{
g1 <- sample(0:2, 10, TRUE)
g2 <- sample(0:2, 10, TRUE)
get.correl.fc(g1, g2)

}
