% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{get.gender}
\alias{get.gender}
\title{Get gender based on heterozygosity}
\usage{
get.gender(sample.summary, threshM, threshF)
}
\arguments{
\item{sample.summary}{Data frame with `Heterozygosity` column.}

\item{threshM}{Numeric threshold for males.}

\item{threshF}{Numeric threshold for females.}
}
\value{
Data frame with columns `heterozygosity` and `sex`.
}
\description{
Infers gender using heterozygosity thresholds.
}
\examples{
df <- data.frame(Heterozygosity = c(0.1, 0.3, 0.6))
rownames(df) <- c("A", "B", "C")
get.gender(df, 0.2, 0.5)

}
