% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{get.hwe.chi2}
\alias{get.hwe.chi2}
\title{Get HWE chi-square p-values}
\usage{
get.hwe.chi2(snp.summary)
}
\arguments{
\item{snp.summary}{Data frame with columns `Calls`, `P.AA`, `P.AB`, `P.BB`.}
}
\value{
Numeric vector with p-values.
}
\description{
Calculates Hardy-Weinberg equilibrium chi-square p-values for SNPs.
}
\examples{
df <- data.frame(Calls = c(100, 100), P.AA = c(0.6, 0.4), P.AB = c(0.3, 0.4), P.BB = c(0.1, 0.2))
get.hwe.chi2(df)

}
\author{
Roberto Higa, adapted by Vinicius Junqueira
}
