% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/haplo_comp.R
\name{haplo_comp}
\alias{haplo_comp}
\title{Function to calculate haplotype frequencies and use
	them to estimate genotype probabilities}
\usage{
haplo_comp(x)
}
\arguments{
\item{x}{A data.frame that contains haplotypes for individuals
(rows) of a population, for a number of SNPs (columns). Haplotypes coded
1 or 2.}
}
\value{
Haplotype frequencies
}
\description{
Function to calculate haplotype frequencies and use
	them to estimate genotype probabilities
}

