% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{ibs.pair}
\alias{ibs.pair}
\title{IBS pair statistics}
\usage{
ibs.pair(g1, g2)
}
\arguments{
\item{g1}{Genotype vector for first sample.}

\item{g2}{Genotype vector for second sample.}
}
\value{
Numeric vector: [mean IBS, standard deviation].
}
\description{
Calculates IBS mean and standard deviation between two samples.
}
\examples{
g1 <- sample(0:2, 10, TRUE)
g2 <- sample(0:2, 10, TRUE)
ibs.pair(g1, g2)

}
