% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_batch.R
\name{import_geno_list}
\alias{import_geno_list}
\title{Import multiple genotype datasets from a list of configurations}
\usage{
import_geno_list(config_list)
}
\arguments{
\item{config_list}{List of configurations (see documentation).}
}
\value{
Combined SNPDataLong object with xref_path slot correctly filled.
}
\description{
Import multiple genotype datasets from a list of configurations
}
