% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeno.R
\name{importAllGenos}
\alias{importAllGenos}
\alias{importAllGenos,SNPImportList-method}
\title{Import and combine multiple genotype configurations}
\usage{
importAllGenos(object)

\S4method{importAllGenos}{SNPImportList}(object)
}
\arguments{
\item{object}{An \code{SNPImportList} object.}
}
\value{
A combined \code{SNPDataLong} object.
}
\description{
Imports genotype data from multiple configurations defined in an
\code{SNPImportList} object and combines them into a unified \code{SNPDataLong} object.
}
