% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFimpute.R
\name{importFImputeResults}
\alias{importFImputeResults}
\title{Import imputed FImpute results from disk}
\usage{
importFImputeResults(path, method = "R")
}
\arguments{
\item{path}{Character. Path to the folder containing 'output_fimpute' (e.g., "fimpute_run_nelore").}

\item{method}{Character. "R" (default) or "Rcpp". Passed to read.fimpute().}
}
\value{
An object of class SNPDataLong containing the imputed genotypes and SNP map.
}
\description{
Reads existing imputed results from a given path and returns an object of class SNPDataLong.
}
