% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_geno.R
\name{merge_geno}
\alias{merge_geno}
\title{Merges SNP array data}
\usage{
merge_geno(...)
}
\arguments{
\item{...}{matrix of SNP array datasets}
}
\value{
matrix of merged SNP array data
}
\description{
Merges any number of genotyping datasets 
and ensures SNPs from each set are merged in the correct order.
}
\details{
Assumes IDs are in rows of the dataset and SNPs are in columns.
}

