% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fQC.R
\name{pairs2sets}
\alias{pairs2sets}
\title{Convert pairs to sets}
\usage{
pairs2sets(pairs)
}
\arguments{
\item{pairs}{Matrix or list of sample pairs.}
}
\value{
List of sets of samples.
}
\description{
Groups sample pairs into sets of related samples.
}
\examples{
pairs <- matrix(c("A", "B", "B", "C", "D", "E"), ncol = 2, byrow = TRUE)
pairs2sets(pairs)

}
\author{
Roberto Higa
}
