% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_anticlustering.R
\name{plotPCAgroups}
\alias{plotPCAgroups}
\title{Plot PCA groups from anticlustering result}
\usage{
plotPCAgroups(pca_res, groups, pcs = c(1, 2), filename = NULL)
}
\arguments{
\item{pca_res}{A prcomp object.}

\item{groups}{A factor or vector of group assignments.}

\item{pcs}{Vector of length 2 indicating which PCs to plot (default: c(1, 2)).}

\item{filename}{Optional. If provided, saves plot to this file (e.g., "antic.png").}
}
\value{
A ggplot object (also prints to screen).
}
\description{
Plot PCA groups from anticlustering result
}
\examples{
\dontrun{
res <- runAnticlusteringPCA(nelore_imputed, K = 2, n_pcs = 20)
plotPCAgroups(res$pca, res$groups)
}

}
