% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_display.R
\name{qc_header}
\alias{qc_header}
\title{Formatted header message}
\usage{
qc_header(title)
}
\arguments{
\item{title}{Character string to be printed inside the header box.}
}
\value{
No return value. Used for side effects (message).
}
\description{
Prints a formatted message with a border for section titles in the console.
}
\examples{
qc_header("Quality Control on Samples")

}
