% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSamples.R
\name{qcSamples}
\alias{qcSamples}
\alias{qcSamples,SNPDataLong-method}
\title{Quality control on samples}
\usage{
qcSamples(x, ...)

\S4method{qcSamples}{SNPDataLong}(
  x,
  heterozygosity = NULL,
  smp_cr = NULL,
  action = c("report", "filter", "both")
)
}
\arguments{
\item{x}{An object of class `SNPDataLong`.}

\item{heterozygosity}{A numeric threshold or range for heterozygosity. Samples outside this threshold are removed.}

\item{smp_cr}{Minimum acceptable sample call rate (between 0 and 1). Samples below this value are removed.}

\item{action}{Character string indicating the action to perform. One of:
- `"report"`: only returns a list of samples to remove and those kept;
- `"filter"`: returns a filtered object without reporting;
- `"both"`: performs filtering and returns the filtered object.}
}
\value{
Depending on the `action` argument:
  - `"report"`: returns a list with removed and kept samples;
  - `"filter"`: returns a new `SNPDataLong` object with filtered genotypes;
  - `"both"`: returns a list with:
      - `filtered`: the filtered `SNPDataLong` object;
      - `report`: a list of removed and kept samples.
}
\description{
Applies quality control (QC) procedures to samples in a `SNPDataLong` object,
based on heterozygosity and call rate thresholds.
}
