% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimpute_runner.R
\name{read.fimpute}
\alias{read.fimpute}
\title{Read imputed genotypes from FImpute output and return SNPDataLong object}
\usage{
read.fimpute(file)
}
\arguments{
\item{file}{Character. Path to the FImpute output directory (usually "output_fimpute").}
}
\value{
An object of class \code{SNPDataLong} containing the imputed genotypes and SNP map.
}
\description{
Reads imputed genotypes and SNP information from FImpute output,
builds a SnpMatrix and a corresponding map, and returns an SNPDataLong object.
}
\examples{
\dontrun{
snp_long <- read.fimpute("output_fimpute")
}
}
