% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recode_geno.R
\name{recode_geno}
\alias{recode_geno}
\title{recode_geno - converts an ACTG coded SNP to a A/B coded SNP}
\usage{
recode_geno(SNPname, geno, table, Bdose = FALSE)
}
\arguments{
\item{SNPname}{character of SNP name to recode}

\item{geno}{matrix of genotyping data with individuals in rows and SNPs in columns.
Formatted in ATCG "forward" format}

\item{table}{data.frame consisting of an equivalency table needed for conversion}

\item{Bdose}{boolean. If true, returns dosage of allele B. If false, returns A/B.}
}
\value{
numeric/character vector with recoded genotypes for all individuals for a single
 SNP in SNPname
}
\description{
recode_geno - converts an ACTG coded SNP to a A/B coded SNP
}

