% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimpute_runner.R
\name{runFImpute}
\alias{runFImpute}
\alias{runFImpute,FImputeRunner-method}
\title{Run FImpute from a FImputeRunner object}
\usage{
runFImpute(object)

\S4method{runFImpute}{FImputeRunner}(object)
}
\arguments{
\item{object}{An object of class `FImputeRunner`}
}
\value{
An updated `FImputeRunner` object with the `results` slot populated.
}
\description{
This function runs the external FImpute software using a `FImputeRunner` object,
ensuring that all required input files are present and the results are imported.
}
