% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fimpute.R
\name{saveFImpute,SNPDataLong-method}
\alias{saveFImpute,SNPDataLong-method}
\title{Exporta genótipos e mapa de um objeto SNPDataLong no formato FImpute}
\usage{
\S4method{saveFImpute}{SNPDataLong}(object, path = "fimpute_run", name = "gen_data")
}
\arguments{
\item{object}{Objeto da classe `SNPDataLong`}

\item{path}{Diretório onde os arquivos serão salvos (default = "fimpute_run")}

\item{name}{Nome base dos arquivos gerados (default = "gen_data")}
}
\value{
NULL (arquivos são salvos no disco)
}
\description{
Esta função cria internamente um objeto da classe FImputeExport e exporta os arquivos `.gen` e `.map`.
}
