% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fimpute.R
\name{saveFImpute}
\alias{saveFImpute}
\alias{saveFImpute,FImputeExport-method}
\alias{saveFImpute,SNPDataLong-method}
\title{Save genotype and map files in FImpute format}
\usage{
saveFImpute(object, ...)

\S4method{saveFImpute}{FImputeExport}(object)

\S4method{saveFImpute}{SNPDataLong}(object, path = NULL)
}
\arguments{
\item{object}{An object of class `FImputeExport` or `SNPDataLong`.}

\item{...}{Additional arguments passed to methods.}

\item{path}{Output directory (default: "fimpute_run" for SNPDataLong).}
}
\value{
No return value. Files are saved to disk.
}
\description{
S4 method to export genotype (.gen), map (.map), and parameter (fimpute.par) files compatible with [FImpute](https://www.aps.uoguelph.ca/~msargol/fimpute/).
}
