% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fimpute.R
\name{saveFImputeRaw}
\alias{saveFImputeRaw}
\title{Exporta genótipos e mapa no formato FImpute a partir de argumentos simples}
\usage{
saveFImputeRaw(geno, map, path, name)
}
\arguments{
\item{geno}{Objeto do tipo SnpMatrix (do pacote snpStats)}

\item{map}{Data frame com colunas 'Name', 'Chromosome' e 'Position'}

\item{path}{Caminho onde os arquivos serão salvos}

\item{name}{Nome base para os arquivos de saída}
}
\description{
Exporta genótipos e mapa no formato FImpute a partir de argumentos simples
}
