% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SNPFileConfig-class}
\alias{SNPFileConfig-class}
\title{SNPFileConfig Class}
\description{
A class for configuring SNP file import options.
}
\section{Slots}{

\describe{
\item{\code{path}}{Path to the SNP file.}

\item{\code{fields}}{A list specifying column mappings or field configurations.}

\item{\code{codes}}{Character vector for genotype or allele codes.}

\item{\code{threshold}}{Numeric value for filtering or quality control.}

\item{\code{sep}}{Character specifying the field separator.}

\item{\code{skip}}{Number of lines to skip at the top of the file.}
}}

