% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SNPImportList-class}
\alias{SNPImportList-class}
\title{SNPImportList Class}
\description{
A class for managing a list of SNP file import configurations.
}
\section{Slots}{

\describe{
\item{\code{configs}}{A list of SNPFileConfig objects.}
}}

