% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPtools-package.R, R/SNPtools-package.R..R
\docType{package}
\name{SNPtools}
\alias{SNPtools}
\title{SNPtools}
\description{
S4 tools for reading and organizing genetic data.

Utilities for reading, cleaning, summarizing, and modeling SNP genotype data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://viniciusjunqueira.github.io/SNPtools}
  \item Report bugs at \url{https://github.com/viniciusjunqueira/SNPtools/issues}
}


Useful links:
\itemize{
  \item \url{https://viniciusjunqueira.github.io/SNPtools}
  \item Report bugs at \url{https://github.com/viniciusjunqueira/SNPtools/issues}
}

}
\author{
Vinícius Junqueira \email{junqueiravinicius@hotmail.com}

\strong{Maintainer}: Vinícius Junqueira \email{junqueiravinicius@hotmail.com}

}
\keyword{internal}
