% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/string_to_vec.R
\name{string_to_vec}
\alias{string_to_vec}
\title{Converts strings to vectors}
\usage{
string_to_vec(char)
}
\arguments{
\item{char}{character string that will be converted to vector}
}
\value{
character vector
}
\description{
Takes in a character string and generates a vector whose elements are
 each a character from the string.
}
\details{
Useful for such things as converting genotype or haplotye data that is
   stored in a string.
}

