% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geno.R
\name{Subset}
\alias{Subset}
\alias{Subset,SNPDataLong-method}
\title{Subset an SNPDataLong object}
\usage{
Subset(object, index, margin = 1, keep = TRUE)

\S4method{Subset}{SNPDataLong}(object, index, margin = 1, keep = TRUE)
}
\arguments{
\item{object}{A \code{SNPDataLong} object.}

\item{index}{Character vector with row (individual) or column (SNP) names to filter.}

\item{margin}{Integer: 1 = rows (individuals), 2 = columns (SNPs).}

\item{keep}{Logical; if \code{TRUE}, keeps the specified names; if \code{FALSE}, removes them.}
}
\value{
A new \code{SNPDataLong} object, subsetted accordingly.
}
\description{
Subsets an \code{SNPDataLong} object by rows (individuals) or columns (SNPs).
You can specify which individuals or SNP markers to keep or remove.
}
