% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,SNPDataLong-method}
\alias{summary,SNPDataLong-method}
\title{Summary para objetos da classe SNPDataLong}
\usage{
\S4method{summary}{SNPDataLong}(object, ...)
}
\arguments{
\item{object}{Objeto do tipo SNPDataLong}
}
\value{
Imprime resumo na tela
}
\description{
Summary para objetos da classe SNPDataLong
}
