% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fimpute_runner.R
\name{FImputeRunner}
\alias{FImputeRunner}
\title{Build FImputeRunner object}
\usage{
FImputeRunner(geno, map, path, exec_path = "FImpute3", name = "data")
}
\arguments{
\item{geno}{A SnpMatrix object.}

\item{map}{A data.frame with SNP metadata (columns: Name, Chromosome, Position).}

\item{path}{A character string indicating the directory to save FImpute files.}

\item{exec_path}{Path to the FImpute executable (default = "FImpute3").}

\item{name}{Name for the dataset (used internally, default = "gen_data").}
}
\value{
An object of class `FImputeRunner`.
}
\description{
A convenience function to construct a `FImputeRunner` object from basic inputs.
}
