% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeno.R
\name{getGeno}
\alias{getGeno}
\title{Leitura flexível e eficiente de genótipos com autodetecção usando fread}
\usage{
getGeno(...)
}
\arguments{
\item{path}{Caminho para o FinalReport.txt}

\item{fields}{Lista com as colunas: sample, snp, allele1, allele2, confidence}

\item{codes}{Vetor com os códigos de alelos (ex: c("A", "B"))}

\item{threshold}{Corte para qualidade (confidence)}

\item{sep}{Separador usado no arquivo}

\item{skip}{Linhas a pular no topo}

\item{verbose}{Exibir progresso?}

\item{every}{Frequência de progresso}
}
\value{
Objeto da classe SNPDataLong ou NULL em caso de erro
}
\description{
Leitura flexível e eficiente de genótipos com autodetecção usando fread
}
