% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_batch.R
\name{import_geno_list}
\alias{import_geno_list}
\title{Importa múltiplos conjuntos genotípicos a partir de uma lista de configurações}
\usage{
import_geno_list(config_list)
}
\arguments{
\item{config_list}{Lista com configurações (cada elemento é uma lista com path, fields, sep, etc.)}
}
\value{
Objeto SNPDataLong unificado
}
\description{
Importa múltiplos conjuntos genotípicos a partir de uma lista de configurações
}
