% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeno.R
\name{importAllGenos}
\alias{importAllGenos}
\title{Importa e combina múltiplas configurações de genótipos}
\usage{
importAllGenos(object)
}
\arguments{
\item{object}{Objeto do tipo SNPImportList}
}
\value{
Objeto SNPDataLong combinado
}
\description{
Importa e combina múltiplas configurações de genótipos
}
