% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeno.R
\name{importAllGenos}
\alias{importAllGenos}
\title{Import and combine multiple genotype configurations}
\usage{
importAllGenos(object)
}
\arguments{
\item{object}{An object of class \code{SNPImportList} containing import configurations}
}
\value{
A single combined \code{SNPDataLong} object
}
\description{
This generic and method import genotype data from multiple configurations defined in an
\code{SNPImportList} object, then combine them into a single unified \code{SNPDataLong} object.
}
