% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimpute_runner.R
\name{importFImputeResults}
\alias{importFImputeResults}
\title{Import imputed FImpute results from disk}
\usage{
importFImputeResults(path)
}
\arguments{
\item{path}{Character. Path to the folder containing 'output_fimpute' (e.g., "fimpute_run_nelore").}
}
\value{
An object of class SNPDataLong containing the imputed genotypes and SNP map.
}
\description{
Reads existing imputed results from a given path and returns an object of class SNPDataLong.
}
\examples{
\dontrun{
imputed_obj <- importFImputeResults("fimpute_run_nelore")
head(imputed_obj@map)
}
}
