% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSNPs.R
\name{qcSNPs}
\alias{qcSNPs}
\title{Controle de Qualidade para SNPDataLong com critérios opcionais}
\usage{
qcSNPs(x, ...)
}
\arguments{
\item{x}{Objeto da classe SNPDataLong}

\item{missing_ind}{Proporção máxima de dados faltantes permitida por indivíduo (opcional)}

\item{missing_snp}{Proporção máxima de dados faltantes permitida por SNP (opcional)}

\item{min_snp_cr}{Mínimo call rate aceitável para SNPs}

\item{min_maf}{Frequência alélica mínima permitida para SNPs (opcional)}

\item{hwe}{Identifica SNPs com desvio de HWE}

\item{snp_mono}{Identifica SNPs monomórficos}

\item{remove_chr}{Descarta SNPs localizados nos cromossomos listados}

\item{action}{"report", "filter" ou "both"}
}
\value{
Dependendo do argumento `action`, retorna:
- "report": lista com SNPs removidos por critério;
- "filter": objeto `SNPDataLong` com SNPs filtrados;
- "both": lista contendo o objeto filtrado e o relatório.
}
\description{
Permite análise de qualidade genotípica com critérios definidos pelo usuário.
}
\examples{
## Not run:
set.seed(123)
mat <- matrix(sample(c(0, 1, 2, NA), 100, replace = TRUE, prob = c(0.4, 0.4, 0.15, 0.05)),
              nrow = 10, ncol = 10)
colnames(mat) <- paste0("snp", 1:10)
rownames(mat) <- paste0("ind", 1:10)
map <- data.frame(Name = colnames(mat), Chrom = 1, Position = 1:10)
x <- new("SNPDataLong", geno = mat, map = map)

qcSNPs(x, min_snp_cr = 0.8, min_maf = 0.05, snp_mono = TRUE, action = "report")
## End(Not run)

}
