% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcSNPs.R
\name{qcSNPs}
\alias{qcSNPs}
\title{Controle de Qualidade para SNPDataLong com critérios opcionais}
\usage{
qcSNPs(x, ...)
}
\arguments{
\item{x}{Objeto da classe SNPDataLong}

\item{missing_ind}{Proporção máxima de dados faltantes permitida por indivíduo (opcional)}

\item{missing_snp}{Proporção máxima de dados faltantes permitida por SNP (opcional)}

\item{min_maf}{Frequência alélica mínima permitida para SNPs (opcional)}

\item{action}{"report", "filter" ou "both"}
}
\value{
Lista com sumário dos filtros e (opcionalmente) os dados filtrados
}
\description{
Permite análise de qualidade genotípica com critérios definidos pelo usuário.
}
