% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimpute_runner.R
\name{runFImpute}
\alias{runFImpute}
\title{Executa o FImpute a partir de um objeto FImputeRunner}
\usage{
runFImpute(object)
}
\arguments{
\item{object}{Objeto da classe FImputeRunner}
}
\value{
Objeto FImputeRunner com o slot `results` preenchido
}
\description{
Esta função executa o software externo FImpute a partir de um objeto da classe FImputeRunner,
garantindo que os arquivos necessários estejam disponíveis e que os resultados possam ser lidos.
}
