% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fimpute.R
\name{saveFImpute}
\alias{saveFImpute}
\title{Método S4 para salvar arquivos no formato FImpute}
\usage{
saveFImpute(object, ...)
}
\arguments{
\item{object}{Objeto da classe FImputeExport}
}
\description{
Método S4 para salvar arquivos no formato FImpute
}
