% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geno.R
\name{Subset}
\alias{Subset}
\title{Subset method for SNPDataLong}
\usage{
Subset(object, index, margin = 1, keep = TRUE)
}
\arguments{
\item{object}{A SNPDataLong object.}

\item{index}{Character vector with row (individual) or column (SNP) names to filter.}

\item{margin}{Integer: 1 = rows (individuals), 2 = columns (SNPs).}

\item{keep}{Logical: TRUE to keep specified levels, FALSE to discard them.}
}
\value{
A new SNPDataLong object, subsetted accordingly.
}
\description{
Subset method for SNPDataLong
}
