% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,SNPDataLong-method}
\alias{summary,SNPDataLong-method}
\title{Summary for SNPDataLong objects}
\usage{
\S4method{summary}{SNPDataLong}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SNPDataLong}.}
}
\value{
Prints a summary to the console. Returns \code{NULL} (invisible).
}
\description{
Provides a detailed summary of an \code{SNPDataLong} object, including sample and SNP counts,
proportion of missing data, and SNP distribution by chromosome if mapping information is available.
}
